+++ components/molecule/modal/src/MoleculeModal.js
@@ -23,6 +23,7 @@ const MoleculeModal = forwardRef(
       children,
       closeOnEscKeyDown = false,
       closeOnOutsideClick = false,
+      isPageScrollable = true,
       enableContentScroll = false,
       fitContent = false,
       floatingIconClose = false,
@@ -60,9 +61,12 @@ const MoleculeModal = forwardRef(
       ev => {
         ev && ev.stopPropagation()
         toggleWindowScroll(false)
+
+        if (!isPageScrollable) document.body.style.overflow = 'auto'
+
         onClose()
       },
-      [onClose]
+      [isPageScrollable, onClose]
     )
 
     const onKeyDown = useCallback(
@@ -75,6 +79,10 @@ const MoleculeModal = forwardRef(
       },
       [isOpen, closeOnEscKeyDown, closeModal]
     )
+
+    useEffect(() => {
+      if (isOpen && !isPageScrollable) document.body.style.overflow = 'hidden'
+    }, [isOpen, isPageScrollable])
 
     useEffect(() => {
       if (usePortal) setIsClientReady(true)
@@ -184,7 +192,11 @@ MoleculeModal.propTypes = {
    */
   children: PropTypes.node,
   /**
-   * true to prevent scroll
+   * false to prevent scroll body
+   */
+  isPageScrollable: PropTypes.node,
+  /**
+   * true to prevent scroll in content
    */
   enableContentScroll: PropTypes.bool,
   /**